Remove Binary Tape flavor and remove game specific flavors
With CK3 1.5, it is now evident that the flavor of the save game can
change from one patch to the next, as the floating format has changed.
This PR allows the binary tape to be constructed in a flavor agnostic
way. That way just a header can be deserialized so that a game version
can be deduced (with a dummy flavor supplied as needed). Then once the
appropriate flavor has been located the data will be deserialized for
real this time without needing to reparse the data.

In short, this PR makes it so one doesn't need to reparse the data when
the binary flavor is unknown.

This PR also removes the game specific flavors: eu4 and ck3. For one, it
was inconsistent that only eu4 and ck3 lived in this repo while
imperator and hoi4 lived in their dedicated save repos. Secondly, now
that a game can have more than one flavor it doesn't make sense to use
their generic game names anymore (`ck3_pre_15` would be more apt).
Third, it's better to track changes to the flavor inside the save repo.

This is a major breaking change.

Instead of:

```rust
BinaryDeserializer::from_eu4(/* ... */)
```

write

```rust
BinaryDeserializer::builder_flavor(InsertFlavorHere)
    .from_slice(/* ... */);
```

`BinaryFlavor` and `Encoding` no longer has the `Sized` bound on the
trait. This follows the [C-OBJECT][0] API guidelines (Traits are
object-safe if they may be useful as a trait object). Binary flavor
methods only receive self and concrete types so removing the `Sized`
bound allows downstream users to put a binary flavor into a trait
object.

[0]: https://rust-lang.github.io/api-guidelines/flexibility.html#c-object
