+++ fuzz/fuzz_targets/fuzz_date.rs
@@ -1,6 +1,6 @@
 #![no_main]
-use libfuzzer_sys::fuzz_target;
 use jomini::common::PdsDate;
+use libfuzzer_sys::fuzz_target;
 
 fuzz_target!(|data: &[u8]| {
     if data.len() < 4 {
