+++ src/binary/flavor.rs
@@ -1,7 +1,8 @@
-use crate::{Encoding, Utf8Encoding, Windows1252Encoding};
-
 /// Trait customizing decoding values from binary data
-pub trait BinaryFlavor: Sized + Encoding {
+///
+/// How binary data is encoded differs between games and even
+/// between patches!
+pub trait BinaryFlavor: crate::Encoding {
     /// Decode a f32 from 4 bytes of data
     fn visit_f32(&self, data: [u8; 4]) -> f32;
 
@@ -9,7 +10,7 @@ pub trait BinaryFlavor: Sized + Encoding {
     fn visit_f64(&self, data: [u8; 8]) -> f64;
 }
 
-impl<T: BinaryFlavor> BinaryFlavor for &'_ T {
+impl<T: BinaryFlavor + ?Sized> BinaryFlavor for &'_ T {
     fn visit_f32(&self, data: [u8; 4]) -> f32 {
         (**self).visit_f32(data)
     }
@@ -19,76 +20,12 @@ impl<T: BinaryFlavor> BinaryFlavor for &'_ T {
     }
 }
 
-/// The eu4 binary flavor
-#[derive(Debug, Default)]
-pub struct Eu4Flavor(Windows1252Encoding);
-
-impl Eu4Flavor {
-    /// Creates a new eu4 flavor
-    pub fn new() -> Self {
-        Eu4Flavor(Windows1252Encoding::new())
-    }
-}
-
-impl Encoding for Eu4Flavor {
-    fn decode<'a>(&self, data: &'a [u8]) -> std::borrow::Cow<'a, str> {
-        self.0.decode(data)
-    }
-}
-
-impl BinaryFlavor for Eu4Flavor {
+impl<T: BinaryFlavor + ?Sized> BinaryFlavor for Box<T> {
     fn visit_f32(&self, data: [u8; 4]) -> f32 {
-        // First encoding is an i32 that has a fixed point offset of 3 decimal digits
-        i32::from_le_bytes(data) as f32 / 1000.0
+        (**self).visit_f32(data)
     }
 
     fn visit_f64(&self, data: [u8; 8]) -> f64 {
-        // Second encoding is Q49.15 with 5 fractional digits
-        // https://en.wikipedia.org/wiki/Q_(number_format)
-        let val = i64::from_le_bytes(data) as f64 / 32768.0;
-        (val * 10_0000.0).round() / 10_0000.0
-    }
-}
-
-/// The ck3 binary flavor
-#[derive(Debug, Default)]
-pub struct Ck3Flavor(Utf8Encoding);
-
-impl Ck3Flavor {
-    /// Creates a new ck3 flavor
-    pub fn new() -> Self {
-        Ck3Flavor(Utf8Encoding::new())
-    }
-}
-
-impl Encoding for Ck3Flavor {
-    fn decode<'a>(&self, data: &'a [u8]) -> std::borrow::Cow<'a, str> {
-        self.0.decode(data)
-    }
-}
-
-impl BinaryFlavor for Ck3Flavor {
-    fn visit_f32(&self, data: [u8; 4]) -> f32 {
-        f32::from_bits(u32::from_le_bytes(data))
-    }
-
-    fn visit_f64(&self, data: [u8; 8]) -> f64 {
-        i64::from_le_bytes(data) as f64 / 1000.0
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    #[test]
-    fn eu4_flavor_f64_rounding() {
-        // This test was taken by running an observer game (plaintext)
-        // with cloud auto save (binary) and comparing the two and
-        // noticing that truncation instead of rounding would yield
-        // `2.49859` instead of the expected `2.49860`
-        let flavor = Eu4Flavor(Windows1252Encoding::new());
-        let data: [u8; 8] = [210, 63, 1, 0, 0, 0, 0, 0];
-        let actual = flavor.visit_f64(data);
-        assert_eq!(actual, 2.49860);
+        (**self).visit_f64(data)
     }
 }
