+++ src/binary/mod.rs
@@ -6,6 +6,6 @@ mod tape;
 
 #[cfg(feature = "derive")]
 pub use self::de::{BinaryDeserializer, BinaryDeserializerBuilder};
-pub use self::flavor::{BinaryFlavor, Ck3Flavor, Eu4Flavor};
+pub use self::flavor::BinaryFlavor;
 pub use self::resolver::{FailedResolveStrategy, TokenResolver};
 pub use self::tape::{BinaryTape, BinaryTapeParser, BinaryToken};
